requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
torch
torchvision
numpy
